requirements.txt
streamlit==0.90.0
tensorflow==2.5.0
numpy==1.19.5
Pillow==8.2.0